menghapus folder duplicate API
