--- app/src/main/java/com/example/tugasweek9/api/Network.kt
@@ -1,57 +1,0 @@
-package com.example.tugasweek9.api
-
-import android.content.Context
-import com.chuckerteam.chucker.api.ChuckerCollector
-import com.chuckerteam.chucker.api.ChuckerInterceptor
-import com.chuckerteam.chucker.api.RetentionManager
-import com.google.gson.GsonBuilder
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-
-object Network {
-
-    private fun builder(context: Context): Retrofit{
-        val chuckerCollector = ChuckerCollector (
-            context = context,
-            showNotification = true,
-            retentionPeriod = RetentionManager.Period.ONE_HOUR
-        )
-
-        val chuckerInterceptor = ChuckerInterceptor.Builder(context)
-            .collector(chuckerCollector)
-            .maxContentLength(250_000L)
-            .createShortcut(true)
-            .build()
-
-        val loggingInterceptor = HttpLoggingInterceptor()
-        loggingInterceptor.level = HttpLoggingInterceptor.Level.BODY
-
-        val client = OkHttpClient.Builder()
-            .connectTimeout(30, TimeUnit.SECONDS)
-            .writeTimeout(30, TimeUnit.SECONDS)
-            .readTimeout(30, TimeUnit.SECONDS)
-            .addInterceptor(HeaderInterceptor())
-            .addInterceptor(loggingInterceptor)
-            .addInterceptor(chuckerInterceptor)
-            .build()
-
-        val gson = GsonBuilder().create()
-
-        return Retrofit.Builder()
-            .baseUrl("https://api.themoviedb.org/3/")
-            .client(client)
-            .addConverterFactory(GsonConverterFactory.create(gson))
-            .build()
-
-
-    }
-
-    fun getService(context: Context): Routes = builder(context).create(Routes::class.java)
-
-
-
-
-}
