+++ app/src/main/java/com/example/tugasweek9/data/api/Routes.kt
@@ -1,6 +1,7 @@
 package com.example.tugasweek9.data.api
 
 import com.example.tugasweek9.data.response.NowPlayingResponse
+import com.example.tugasweek9.response.UpcomingResponse
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Query
@@ -18,4 +19,10 @@ interface Routes {
         @Query("language") lang: String = "en-US",
         @Query("page") page: Int
     ): NowPlayingResponse
+
+    @GET("movie/upcoming")
+    suspend fun getUpcoming(
+        @Query("language") lang: String = "en-EN",
+        @Query("page") page: Int
+    ): UpcomingResponse
 }
